requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
openpyxl>=3.1.0
XlsxWriter>=3.0.0
psycopg2-binary>=2.9.0
bcrypt>=4.0.0
sqlalchemy>=2.0.0
folium>=0.14.0
streamlit-folium>=0.17.0
streamlit-autorefresh>=1.0.0
Pillow>=10.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0
streamlit-autorefresh
fastapi>=0.104.0
uvicorn>=0.24.0
python-jose[cryptography]>=3.5.0
passlib[bcrypt]>=1.7.0
python-multipart>=0.0.6
pyinstaller>=6.0.0
python-dotenv>=1.0.0